[Update] Implement foreign keys and clean up directory
